add hard links to build.sh in precise/trusty dirs
In order to setup automatic builds on hub.docker.com the entire context
of the build needs to be available in the same directory as the
Dockerfile and you can't easily specify a Dockerfile other than saying
what directory it is in. The last option is to maintain the file shared
files in multiple places, but with a link that means they all get
updated at the same time. Yay.